Un-nest App from the the router components.
